F411758:US000000-update PFNY regression script

+++ cukesatdd/src/main/java/pages/regression/prepareForNextYear/PrepareForNextYearBase.java
@@ -42,17 +42,13 @@ import acceptancetests.util.CommonUtility;
 import atdd.framework.MRScenario;
 
 public class PrepareForNextYearBase  extends PrepareForNextYearWebElements {
-	protected boolean validateAsMuchAsPossible;
+	protected static boolean validateAsMuchAsPossible=true;
 	
 	protected static final String m1="09/15/";
 	protected static final String m2="10/01/";
 	protected static final String m3="10/15/";
 	protected static final String m4="12/07/";
 	
-	public void setValidateAsMuchAsPossible(boolean input_validateAsMuchAsPossible) {
-		validateAsMuchAsPossible=input_validateAsMuchAsPossible;
-	}
-
 	public PrepareForNextYearBase(WebDriver driver) {
 		super(driver);
 		PageFactory.initElements(driver, this);
@@ -535,6 +531,8 @@ public class PrepareForNextYearBase  extends PrepareForNextYearWebElements {
 
 	public List<String> validateDontHaveItem(String targetItem, WebElement targetElement) {
 		List<String> note=new ArrayList<String>();
+		System.out.println("TEST - check validateAsMuchAsPossible="+validateAsMuchAsPossible);
+
 		if (validateAsMuchAsPossible) {
 			if (!noWaitValidate(targetElement)) 
 				note.add("\tPASSED - validation for NOT HAVING "+targetItem);
@@ -624,16 +622,32 @@ public class PrepareForNextYearBase  extends PrepareForNextYearWebElements {
 		//tbd 		|| targetItem.contains("Drug Search link")
 		//tbd 		|| targetItem.contains("Find a Pharmacy link")) {
 			String winHandleBefore = driver.getWindowHandle();
-
+			String urlBeforeClick=driver.getCurrentUrl();
 			ArrayList<String> beforeClicked_tabs = new ArrayList<String>(driver.getWindowHandles());
 			int beforeClicked_numTabs=beforeClicked_tabs.size();	
 			CommonUtility.waitForPageLoad(driver, expElement, 5);
+			scrollElementToCenterScreen(targetElement);
 			targetElement.click();
-			CommonUtility.checkPageIsReady(driver);
 			
 			ArrayList<String> afterClicked_tabs = new ArrayList<String>(driver.getWindowHandles());
 			int afterClicked_numTabs=afterClicked_tabs.size();
-			Assert.assertTrue("PROBLEM - Did not get expected new tab after clicking '"+targetItem+"' link", (afterClicked_numTabs-beforeClicked_numTabs)==1);
+			if (validateAsMuchAsPossible) {
+				if ((afterClicked_numTabs-beforeClicked_numTabs)!=1) {
+					note.add("\t * FAILED - Did not get expected new tab after clicking '"+targetItem+"' link. Number of existing tab before link click='"+beforeClicked_numTabs+"' | After='"+afterClicked_numTabs+"'");
+					//note: back to prior page and move on
+					if (!driver.getCurrentUrl().contains("preparefornextyear/overview.html")) {
+						driver.get(urlBeforeClick);
+						goToSpecificComboTab(planType,false);
+						CommonUtility.checkPageIsReady(driver);
+						checkModelPopup(driver,5);
+					}
+					return note;
+				} else {
+					System.out.println("TEST - link opened on new tab as expected...moving on...");
+				}
+			} else {
+					Assert.assertTrue("PROBLEM - Did not get expected new tab after clicking '"+targetItem+"' link. Number of existing tab before link click='"+beforeClicked_numTabs+"' | After='"+afterClicked_numTabs+"'", (afterClicked_numTabs-beforeClicked_numTabs)==1);
+			}
 			driver.switchTo().window(afterClicked_tabs.get(afterClicked_numTabs-1));
 			CommonUtility.checkPageIsReady(driver);
 			CommonUtility.waitForPageLoad(driver, expElement, 10);
@@ -659,7 +673,9 @@ public class PrepareForNextYearBase  extends PrepareForNextYearWebElements {
 			}
 
 			driver.close();
+			System.out.println("TEST - Closed tab for '"+targetItem+"'");
 			driver.switchTo().window(winHandleBefore);
+			System.out.println("TEST - Switched back to prior tab");
 			/*
 		} else { 
 			String originalUrl=driver.getCurrentUrl();
@@ -723,23 +739,34 @@ public class PrepareForNextYearBase  extends PrepareForNextYearWebElements {
 			return true;
 	}	
 	
-	public void selectValueFromDropdown(WebElement dropdownElement, String targetLang) {
-		String value="en_us";
-		if (targetLang.equals("Spanish"))
-			value="es";
-		else if (targetLang.equals("Chinese"))
-			value="zh";
-		
-		scrollElementToCenterScreen(dropdownElement);
-		Select select = new Select(dropdownElement);    
-		waitTillElementClickableInTime(dropdownElement,10);
+	public boolean selectValueFromDropdown(WebElement dropdownElement, String targetLang) {
 		try {
-			dropdownElement.click();
-		} catch (WebDriverException e) { //note: in case it's timing, try one more time before giving up
-			sleepBySec(1);
-			dropdownElement.click();
+			String value="en_us";
+			if (targetLang.equals("Spanish"))
+				value="es";
+			else if (targetLang.equals("Chinese"))
+				value="zh";
+			
+			scrollElementToCenterScreen(dropdownElement);
+			Select select = new Select(dropdownElement);    
+			waitTillElementClickableInTime(dropdownElement,10);
+			try {
+				dropdownElement.click();
+			} catch (WebDriverException e) { //note: in case it's timing, try one more time before giving up
+				sleepBySec(1);
+				dropdownElement.click();
+			}
+			select.selectByValue(value);
+		} catch (WebDriverException e) {
+			if (validateAsMuchAsPossible) {
+				System.out.println("unable to select expected langauge "+targetLang+" option from dropdown");
+				return false;
+			} else {
+				Assert.assertTrue("PROBLEM - unable to select expected langauge "+targetLang+" option from dropdown", false);
+				return false;
+			}
 		}
-		select.selectByValue(value);
+		return true;
 	}
 
 	
@@ -764,48 +791,54 @@ public class PrepareForNextYearBase  extends PrepareForNextYearWebElements {
 		String targetItem=section+" - "+targetLang+" '"+docName+" (PDF)'";
 		if (docDisplayMap.get(docName+" "+targetLang)) {
 			note.add("\tEXPECT "+targetLang+" '"+docName+"' document to display");
-			selectValueFromDropdown(langDropdownElement1, targetLang);
+			if (selectValueFromDropdown(langDropdownElement1, targetLang)) {
+				if (langDropdownElement2!=null) {
+					Select select2 = new Select(langDropdownElement2);           
+					String otherDropDownSelectedValue=select2.getFirstSelectedOption().getText();
+					if (validateAsMuchAsPossible) {
+						if (!otherDropDownSelectedValue.equalsIgnoreCase("ENGLISH")) 
+							note.add("\t * FAILED - switching language option in one section should not have impacted the langage option in other section");
+					} else {
+						Assert.assertTrue("PROBLEM - switching language option in one section should not have impacted the langage option in other section", otherDropDownSelectedValue.equalsIgnoreCase("ENGLISH"));
+					}
+				}
 
-			Select select2 = new Select(langDropdownElement2);           
-			String otherDropDownSelectedValue=select2.getFirstSelectedOption().getText();
-			if (validateAsMuchAsPossible) {
-				if (!otherDropDownSelectedValue.equalsIgnoreCase("ENGLISH")) 
-					note.add("\t * FAILED - switching language option in one section should not have impacted the langage option in other section");
+				CommonUtility.waitForPageLoad(driver, pdfElement, 10);
+				note.addAll(validateHaveItem(targetItem, pdfElement));
+				if (noWaitValidate(pdfElement)) {
+					note.addAll(validatePdfLinkTxt(docName, pdfElement));
+					note.addAll(validatePdf(targetItem, pdfElement));
+				}
+
+				//targetItem=section+" - arrow before pdf";
+				//note.addAll(validateHaveItem(targetItem, arrowBefPdfElement));
+
+				//targetItem=section+" - the 'or' text";
+				//note.addAll(validateHaveItem(targetItem, orTextBefPdfElement));
+
+				targetItem=section+" - Arrow after '"+docName+"' doc link'";
+				note.addAll(validateHaveItem(targetItem, arrowAftPdfElement));
+
+				//note: after link click, little check should turn green
+				//note: some section has inconsistent way to locate the green chkmrk xpath...that's why need to figure out which xpath to use
+				note.add("\n\tValidate after clicking "+targetLang+" '"+docName+"' link");
+				targetItem=section+" - green checkmark";
+				WebElement subSecChkmrkgreen=subSecChkmrkgreen1;
+				if (noWaitValidate(subSecChkmrkgreen1)) {
+					subSecChkmrkgreen=subSecChkmrkgreen1;
+				} else {
+					subSecChkmrkgreen=subSecChkmrkgreen2;
+				}
+				note.addAll(validateHaveItem(targetItem, subSecChkmrkgreen));
+
+				if (willDeleteCookie) {
+					note.add("\n\tValidate after cookie remove for '"+subSection+"' subsection cookie");
+					deleteCookieAndReloadPgn(subSecCookie);
+					note.addAll(validateDontHaveItem(targetItem, subSecChkmrkgreen));
+				}
+
 			} else {
-				Assert.assertTrue("PROBLEM - switching language option in one section should not have impacted the langage option in other section", otherDropDownSelectedValue.equalsIgnoreCase("ENGLISH"));
-			}
-
-			CommonUtility.waitForPageLoad(driver, pdfElement, 10);
-			note.addAll(validateHaveItem(targetItem, pdfElement));
-			note.addAll(validatePdfLinkTxt(docName, pdfElement));
-			note.addAll(validatePdf(targetItem, pdfElement));
-
-
-			//targetItem=section+" - arrow before pdf";
-			//note.addAll(validateHaveItem(targetItem, arrowBefPdfElement));
-
-			//targetItem=section+" - the 'or' text";
-			//note.addAll(validateHaveItem(targetItem, orTextBefPdfElement));
-
-			targetItem=section+" - Arrow after '"+docName+"' doc link'";
-			note.addAll(validateHaveItem(targetItem, arrowAftPdfElement));
-
-			//note: after link click, little check should turn green
-			//note: some section has inconsistent way to locate the green chkmrk xpath...that's why need to figure out which xpath to use
-			note.add("\n\tValidate after clicking "+targetLang+" '"+docName+"' link");
-			targetItem=section+" - green checkmark";
-			WebElement subSecChkmrkgreen=subSecChkmrkgreen1;
-			if (noWaitValidate(subSecChkmrkgreen1)) {
-				subSecChkmrkgreen=subSecChkmrkgreen1;
-			} else {
-				subSecChkmrkgreen=subSecChkmrkgreen2;
-			}
-			note.addAll(validateHaveItem(targetItem, subSecChkmrkgreen));
-
-			if (willDeleteCookie) {
-				note.add("\n\tValidate after cookie remove for '"+subSection+"' subsection cookie");
-				deleteCookieAndReloadPgn(subSecCookie);
-				note.addAll(validateDontHaveItem(targetItem, subSecChkmrkgreen));
+				note.add("\t * FAILED - unable to select expected langauge "+targetLang+" option from dropdown");
 			}
 
 		} else {
